# public/openapi/openapi.yaml
openapi: 3.2.0
info:
    title: 'Rotor API'
    version: 1.0.0
    description: |
        Для доступа к данным нужен API-ключ, который можно получить через авторизацию или на странице Мои данные.

        Лимиты запросов: 120 в минуту на все методы, 10 в минуту на авторизацию.
        Остаток лимита — в заголовках X-RateLimit-Limit и X-RateLimit-Remaining, при превышении приходит 429 и Retry-After.
security:
    -
        BearerToken: {  }
components:
    securitySchemes:
        BearerToken:
            type: http
            scheme: bearer
            description: 'Передайте API-ключ в заголовке: Authorization: Bearer {token}'
    parameters:
...
        PerPage:
            name: per_page
            in: query
            schema:
                type: integer
                default: 10
                minimum: 1
                maximum: 100
            description: 'Количество элементов на странице'
        Page:
            name: page
            in: query
            schema:
                type: integer
                default: 1
                minimum: 1
            description: 'Номер страницы'
        Order:
            name: order
            in: query
            schema:
                type: string
                enum:
                    - asc
                    - desc
                default: desc
            description: 'Порядок сортировки'
    schemas:
...
        SearchItem:
            type: object
            properties:
                type:
                    type: string
                    description: 'Тип найденной записи (comments, users, news, topics и т.д.)'
                id:
                    type: integer
                section:
                    type:
                        - string
                        - 'null'
                    description: 'Название раздела'
                title:
                    type:
                        - string
                        - 'null'
                    description: 'Заголовок записи; у пользователя — логин, у комментария — заголовок записи, к которой он оставлен'
                url:
                    type:
                        - string
                        - 'null'
                    format: uri
                breadcrumbs:
                    type: array
                    description: 'Путь до записи от корня раздела'
                    items:
                        type: object
                        properties:
                            title:
                                type: string
                            url:
                                type: string
                                format: uri
                text:
                    type:
                        - string
                        - 'null'
                    description: 'HTML текста записи'
                user:
                    oneOf:
                        -
                            $ref: '#/components/schemas/Author'
                        -
                            type: 'null'
                created_at:
                    type: string
                    format: date-time
        FeedItem:
            type: object
            properties:
                type:
                    type: string
                    description: 'Тип записи (comments, news, topics, articles и т.д.)'
                id:
                    type: integer
                section:
                    type:
                        - string
                        - 'null'
                    description: 'Название раздела'
                title:
                    type:
                        - string
                        - 'null'
                url:
                    type:
                        - string
                        - 'null'
                    format: uri
                breadcrumbs:
                    type: array
                    description: 'Путь до записи от корня раздела, пустой массив — раздел не указан'
                    items:
                        type: object
                        properties:
                            title:
                                type: string
                            url:
                                type: string
                                format: uri
                text:
                    type:
                        - string
                        - 'null'
                    description: 'HTML текста записи'
                rating:
                    type:
                        - integer
                        - 'null'
                vote:
                    type:
                        - object
                        - 'null'
                    description: 'Цель голосования, null — тип не поддерживает рейтинг'
                    properties:
                        type:
                            type: string
                            description: 'Тип для POST /rating (у темы форума — последнее сообщение)'
                        id:
                            type: integer
                        value:
                            type:
                                - string
                                - 'null'
                            enum:
                                - +
                                - '-'
                                - null
                            description: 'Голос текущего пользователя, null без токена или если не голосовал'
                        own:
                            type: boolean
                            description: 'Своя запись — голосовать нельзя'
                comments_count:
                    type:
                        - integer
                        - 'null'
                user:
                    $ref: '#/components/schemas/Author'
                media:
                    type: array
                    description: 'Картинки и видео для галереи, кроме вставленных в текст'
                    items:
                        $ref: '#/components/schemas/File'
                files:
                    type: array
                    description: 'Остальные вложения — документы, архивы, аудио'
                    items:
                        $ref: '#/components/schemas/File'
                created_at:
                    type: string
                    format: date-time
                relate:
                    type: object
                    description: 'Только при type=comments: запись, к которой оставлен комментарий'
                    properties:
                        type:
                            type: string
                        id:
                            type: integer
                        title:
                            type:
                                - string
                                - 'null'
                        url:
                            type:
                                - string
                                - 'null'
                            format: uri
                price:
                    type:
                        - number
                        - 'null'
                    description: 'Только при type=items (объявления): цена, валюта — в настройке currency'
                phone:
                    type:
                        - string
                        - 'null'
                    description: 'Только при type=items (объявления): телефон продавца'
                offer_type:
                    type: string
                    enum:
                        - offer
                        - issue
                    description: 'Только при type=offers: предложение или проблема'
                status:
                    type: string
                    enum:
                        - wait
                        - process
                        - done
                        - cancel
                    description: 'Только при type=offers: статус обработки'
            additionalProperties: true
            description: |
                Набор полей зависит от type: базовые поля приходят всегда, дополнительные — только у своего типа записи
                (relate у комментариев, price и phone у объявлений, offer_type и status у предложений).
                Сторонний модуль может добавить собственные поля, поэтому неизвестные ключи игнорируйте, а не считайте ошибкой.
...
        SimplePaginationLinks:
            type: object
            description: 'last всегда null — общее число страниц не вычисляется'
            properties:
                first:
                    type: string
                    format: uri
                last:
                    type: 'null'
                prev:
                    type:
                        - string
                        - 'null'
                    format: uri
                next:
                    type:
                        - string
                        - 'null'
                    format: uri
        SimplePaginationMeta:
            type: object
            description: 'Упрощённая пагинация: общее число записей не считается, следующая страница определяется по links.next'
            properties:
                current_page:
                    type: integer
                current_page_url:
                    type: string
                    format: uri
                from:
                    type:
                        - integer
                        - 'null'
                path:
                    type: string
                    format: uri
                per_page:
                    type: integer
                to:
                    type:
                        - integer
                        - 'null'
...
paths:
...
    /feed:
        get:
            tags:
                - Лента
            summary: 'Лента последних записей сайта'
            description: 'Токен не обязателен. С токеном в каждом элементе приходит голос текущего пользователя (vote.value)'
            operationId: getFeed
            security:
                -
                    BearerToken: {  }
                - {}
            parameters:
                -
                    $ref: '#/components/parameters/Page'
            responses:
                200:
                    description: OK
                    content:
                        application/json:
                            schema:
                                type: object
                                properties:
                                    data:
                                        type: array
                                        items:
                                            $ref: '#/components/schemas/FeedItem'
                                    links:
                                        $ref: '#/components/schemas/SimplePaginationLinks'
                                    meta:
                                        $ref: '#/components/schemas/SimplePaginationMeta'
                429:
                    description: 'Превышен общий лимит запросов'
    /search:
        get:
            tags:
                - Поиск
            summary: 'Полнотекстовый поиск по сайту'
            description: |
                Ищет по записям разделов, комментариям и пользователям. Токен не нужен — выдача не зависит от пользователя.

                Слова короче трёх символов отбрасываются, из оставшихся собирается запрос;
                если после отсева ничего не осталось — 422.
            operationId: getSearch
            security: []
            parameters:
                -
                    name: query
                    in: query
                    required: true
                    schema:
                        type: string
                        minLength: 3
                        maxLength: 64
                    description: 'Поисковый запрос (синоним — q)'
                -
                    name: type
                    in: query
                    schema:
                        type: string
                    description: 'Ограничить поиск разделом; допустимые значения приходят в types, неизвестное значение игнорируется'
                -
                    name: sort
                    in: query
                    schema:
                        type: string
                        enum:
                            - relevance
                            - date
                            - date_asc
                        default: relevance
                    description: 'Сортировка выдачи'
                -
                    $ref: '#/components/parameters/PerPage'
                -
                    $ref: '#/components/parameters/Page'
            responses:
                200:
                    description: OK
                    content:
                        application/json:
                            schema:
                                type: object
                                properties:
                                    data:
                                        type: array
                                        items:
                                            $ref: '#/components/schemas/SearchItem'
                                    links:
                                        $ref: '#/components/schemas/PaginationLinks'
                                    meta:
                                        $ref: '#/components/schemas/PaginationMeta'
                                    query:
                                        type: string
                                        description: 'Запрос после очистки от пунктуации'
                                    type:
                                        type:
                                            - string
                                            - 'null'
                                    sort:
                                        type: string
                                    types:
                                        type: object
                                        description: 'Разделы, доступные для фильтра: ключ — значение параметра type, значение — название раздела'
                                        additionalProperties:
                                            type: string
                422:
                    description: 'Слишком короткий или слишком длинный запрос'
                429:
                    description: 'Превышен общий лимит запросов'
...
    /files:
        get:
            tags:
                - Вложения
            summary: 'Свои вложения'
            description: 'Возвращает файлы текущего пользователя. Без id — загруженные, но ещё не привязанные к записи. Требует токен'
            operationId: getFiles
            parameters:
                -
                    name: type
                    in: query
                    required: true
                    schema:
                        type: string
                    description: 'Тип записи (comments, messages, offers, photos, downs)'
                -
                    name: id
                    in: query
                    schema:
                        type: integer
                        default: 0
                    description: 'Id записи; 0 — файлы, ожидающие привязки'
            responses:
                200:
                    description: OK
                    content:
                        application/json:
                            schema:
                                type: object
                                properties:
                                    data:
                                        type: array
                                        items:
                                            $ref: '#/components/schemas/File'
                400:
                    description: 'Токен не передан'
                401:
                    description: 'Неверный токен'
                422:
                    description: 'Неизвестный тип'
                429:
                    description: 'Превышен общий лимит запросов'
        post:
            tags:
                - Вложения
            summary: 'Загрузка вложения'
            description: |
                Требует токен. Ограничения на размер, количество и расширения приходят в /config.

                id = 0 — запись ещё не создана: файл повиснет за пользователем и привяжется при её сохранении.
            operationId: uploadFile
            requestBody:
                required: true
                content:
                    multipart/form-data:
                        schema:
                            type: object
                            required:
                                - type
                                - file
                            properties:
                                type:
                                    type: string
                                    description: 'Тип записи (comments, messages, offers, photos, downs)'
                                id:
                                    type: integer
                                    default: 0
                                file:
                                    type: string
                                    format: binary
            responses:
                201:
                    description: 'Файл загружен'
                    content:
                        application/json:
                            schema:
                                type: object
                                properties:
                                    message:
                                        type: string
                                    file:
                                        $ref: '#/components/schemas/File'
                400:
                    description: 'Токен не передан'
                401:
                    description: 'Неверный токен'
                422:
                    description: 'Неизвестный тип, превышен лимит файлов, дубликат или файл не прошёл проверку'
                429:
                    description: 'Превышен общий лимит запросов'
...
    /user:
        get:
            tags:
                - Пользователи
            summary: 'Профиль текущего пользователя'
            operationId: getUser
            responses:
                200:
                    description: OK
                    content:
                        application/json:
                            schema:
                                type: object
                                properties:
                                    data: { $ref: '#/components/schemas/UserProfile' }
                400:
                    description: 'Api token missing'
                401:
                    description: Unauthorized
    '/users/{login}':
        get:
            tags:
                - Пользователи
            summary: 'Профиль пользователя по логину'
            operationId: getUserByLogin
            parameters:
                -
                    name: login
                    in: path
                    required: true
                    schema:
                        type: string
                        pattern: '^[\w\-]+$'
            responses:
                200:
                    description: OK
                    content:
                        application/json:
                            schema:
                                type: object
                                properties:
                                    data: { $ref: '#/components/schemas/User' }
                404:
                    description: 'Пользователь не найден'
...
    '/talk/{login}':
        get:
            tags:
                - Сообщения
            summary: 'Переписка с пользователем'
            operationId: getTalk
            parameters:
                -
                    name: login
                    in: path
                    required: true
                    schema:
                        type: string
                        pattern: '^[\w\-]+$'
                -
                    $ref: '#/components/parameters/PerPage'
                -
                    $ref: '#/components/parameters/Page'
                -
                    $ref: '#/components/parameters/Order'
            responses:
                200:
                    description: OK
                    content:
                        application/json:
                            schema:
                                type: object
                                properties:
                                    data: { type: array, items: { $ref: '#/components/schemas/Message' } }
                                    links: { $ref: '#/components/schemas/PaginationLinks' }
                                    meta: { $ref: '#/components/schemas/PaginationMeta' }
                                    dialogue: { $ref: '#/components/schemas/Dialogue' }
                404:
                    description: 'Пользователь не найден'
        post:
            tags:
                - Сообщения
            summary: 'Отправить сообщение пользователю'
            operationId: createTalk
            parameters:
                -
                    name: login
                    in: path
                    required: true
                    schema:
                        type: string
                        pattern: '^[\w\-]+$'
            requestBody:
                required: true
                content:
                    multipart/form-data:
                        schema:
                            type: object
                            required:
                                - text
                            properties:
                                text:
                                    type: string
                                    description: 'Текст сообщения'
                                'files[]':
                                    type: array
                                    items: { type: string, format: binary }
                                    description: 'Прикреплённые файлы (необязательно)'
            responses:
                201:
                    description: 'Сообщение отправлено'
                    content:
                        application/json:
                            schema:
                                type: object
                                properties:
                                    message: { type: string }
                                    data: { $ref: '#/components/schemas/Message' }
                404:
                    description: 'Пользователь не найден'
                422:
                    description: 'Ошибка валидации (флуд, игнор-лист, нельзя себе)'
...
    /items:
        get:
            tags:
                - Объявления
            summary: 'Список объявлений'
            description: 'Истёкшие объявления в список не попадают; без category_id — из всех категорий'
            operationId: getItems
            security:
                -
                    BearerToken: {  }
                - {}
            parameters:
                -
                    $ref: '#/components/parameters/CategoryId'
                -
                    $ref: '#/components/parameters/Sort'
                -
                    $ref: '#/components/parameters/PerPage'
                -
                    $ref: '#/components/parameters/Page'
                -
                    $ref: '#/components/parameters/Order'
            responses:
                200:
                    description: OK
                    content:
                        application/json:
                            schema:
                                type: object
                                properties:
                                    data: { type: array, items: { $ref: '#/components/schemas/Item' } }
                                    links: { $ref: '#/components/schemas/PaginationLinks' }
                                    meta: { $ref: '#/components/schemas/PaginationMeta' }
...